refactor: 테마 제거 시 예약에서 참조 중인 테마인 경우 예외가 발생하도록 변경

+++ src/main/java/roomescape/theme/service/ThemeService.java
@@ -3,6 +3,9 @@ package roomescape.theme.service;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.global_exception.BadRequestException;
+import roomescape.global_exception.NotFoundException;
+import roomescape.reservation.repository.ReservationRepository;
 import roomescape.theme.domain.Theme;
 import roomescape.theme.domain.dto.ThemeReqDto;
 import roomescape.theme.domain.dto.ThemeResDto;
@@ -14,9 +17,11 @@ public class ThemeService {
     private static final int BETWEEN_DAY_START = 7;
     private static final int BETWEEN_DAY_END = 1;
     private final ThemeRepository repository;
+    private final ReservationRepository reservationRepository;
 
-    public ThemeService(ThemeRepository repository) {
+    public ThemeService(ThemeRepository repository, ReservationRepository reservationRepository) {
         this.repository = repository;
+        this.reservationRepository = reservationRepository;
     }
 
     public List<ThemeResDto> findAll() {
@@ -41,6 +46,11 @@ public class ThemeService {
     }
 
     public void deleteById(Long id) {
+        Theme theme = repository.findById(id)
+            .orElseThrow(() -> new NotFoundException("존재하지 않는 테마입니다."));
+        if (reservationRepository.existsByTheme(theme)) {
+            throw new BadRequestException("예약에서 사용 중인 테마입니다.");
+        }
         repository.delete(id);
     }
 }
